[#176455937] move screen file outside details dir
